+++ src/test/java/io/github/co_mmer/aaamockmvc/ej/test/web/answer/TestAnswerImplTest.java
@@ -1,6 +1,5 @@
 package io.github.co_mmer.aaamockmvc.ej.test.web.answer;
 
-import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestBody.TEST_BODY_JSON;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestDataRequestDto.TEST_REQUEST_DTO;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestHeader.TEST_HEADER_KEY_1;
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestHeader.TEST_HEADER_VALUE_1;
@@ -15,8 +14,8 @@ import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestObject.TEST_
 import static io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestValue.TEST_BYTE;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.nullValue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -27,374 +26,388 @@ import io.github.co_mmer.aaamockmvc.ej.test.web.answer.exception.TestAnswerExcep
 import io.github.co_mmer.aaamockmvc.ej.test.web.request.context.TestRequestBean;
 import io.github.co_mmer.aaamockmvc.ej.test.web.request.context.TestRequestContext;
 import io.github.co_mmer.aaamockmvc.ej.testdata.MockTestRequestStrategyFactory;
-import io.github.co_mmer.aaamockmvc.ej.testdata.testsetup.MockMvcSetup;
 import io.github.co_mmer.aaamockmvc.ej.testdata.testutil.TestObjectSimple;
-import java.util.stream.Stream;
+import java.io.UnsupportedEncodingException;
+import lombok.SneakyThrows;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Mockito;
+import org.springframework.mock.web.MockHttpServletResponse;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 
 class TestAnswerImplTest {
 
-  private MockMvc mvc;
-  private MockMvcSetup setup;
   private TestAnswer testAnswer;
+  private ResultActions mockResultActions;
+  private TestRequestBean testRequestBean;
   private MockHttpServletRequestBuilder mockRequestBuilder;
-  private TestRequestContext testRequestContext;
 
   @BeforeEach
+  @SneakyThrows
   void setUp() {
-    this.mvc = mock(MockMvc.class);
-    this.setup = new MockMvcSetup(this.mvc);
-    var testRequestConfig = new TestRequestBean(this.mvc, new ObjectMapper());
+    var mvc = mock(MockMvc.class);
     this.mockRequestBuilder = mock(MockHttpServletRequestBuilder.class);
+    this.testRequestBean = new TestRequestBean(mvc, new ObjectMapper());
+    this.mockResultActions = mock(ResultActions.class);
+    when(mvc.perform(this.mockRequestBuilder)).thenReturn(this.mockResultActions);
 
-    this.testRequestContext = new TestRequestContext(TEST_REQUEST_DTO, testRequestConfig);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
+    var testRequestContext = new TestRequestContext(TEST_REQUEST_DTO, testRequestBean);
+    this.testAnswer = new TestAnswerImpl(testRequestContext, this.mockRequestBuilder);
   }
 
-  @Test
-  void GIVEN_exception_WHEN_answerAsResultActions_THEN_throw_exception() throws Exception {
-    // Arrange
-    when(this.mvc.perform(any())).thenThrow(Exception.class);
+  private void useContentAsString(String content) throws Exception {
+    var mockMvcResult = mock(MvcResult.class);
+    var mockMockHttpServletResponse = mock(MockHttpServletResponse.class);
+    when(mockMockHttpServletResponse.getContentAsString()).thenReturn(content);
 
-    // Act
-    var exception = assertThrows(Exception.class, () -> this.testAnswer.answerAsResultActions());
-
-    // Assert
-    assertThat(exception.getClass(), is(TestAnswerException.class));
+    when(mockMvcResult.getResponse()).thenReturn(mockMockHttpServletResponse);
+    when(mockResultActions.andReturn()).thenReturn(mockMvcResult);
   }
 
-  @Test
-  void WHEN_answerVoid_THEN_mvc_perform_is_called() throws Exception {
-    // Arrange
-    var mockTestRequestStrategy = Mockito.mock(TestRequestBaseStrategy.class);
-    var mockTestRequestStrategyFactory =
-        MockTestRequestStrategyFactory.mockTestRequestStrategyFactory(mockTestRequestStrategy);
+  private void useContentAsByte() {
+    var mockMvcResult = mock(MvcResult.class);
+    var mockMockHttpServletResponse = mock(MockHttpServletResponse.class);
+    when(mockMockHttpServletResponse.getContentAsByteArray()).thenReturn(TEST_BYTE);
 
-    when(mockTestRequestStrategy.apply(TEST_REQUEST_DTO)).thenReturn(this.mockRequestBuilder);
-
-    // Act
-    this.testAnswer.answerVoid();
-
-    // Assert
-    verify(this.mvc).perform(this.mockRequestBuilder);
-    mockTestRequestStrategyFactory.close();
+    when(mockMvcResult.getResponse()).thenReturn(mockMockHttpServletResponse);
+    when(mockResultActions.andReturn()).thenReturn(mockMvcResult);
   }
 
-  @Test
-  void GIVEN_exception_WHEN_answerVoid_THEN_throw_TestAnswerException() throws Exception {
-    // Arrange
-    when(this.mvc.perform(any())).thenThrow(Exception.class);
+  private void useContentAsStringException() throws Exception {
+    var mockMvcResult = mock(MvcResult.class);
+    var mockMockHttpServletResponse = mock(MockHttpServletResponse.class);
+    when(mockMockHttpServletResponse.getContentAsString())
+        .thenThrow(new UnsupportedEncodingException("test"));
 
-    // Act
-    var exception = assertThrows(Exception.class, () -> this.testAnswer.answerVoid());
-
-    // Assert
-    assertThat(exception.getClass(), is(TestAnswerException.class));
+    when(mockMvcResult.getResponse()).thenReturn(mockMockHttpServletResponse);
+    when(mockResultActions.andReturn()).thenReturn(mockMvcResult);
   }
 
-  @Test
-  void WHEN_answerAsString_THEN_mvc_getContentAsString_is_called() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString();
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Act
-    this.testAnswer.answerAsString();
-
-    // Assert
-    this.setup.verifyGetContentAsString();
+  private void useHeader() {
+    var mockMvcResult = mock(MvcResult.class);
+    var mockMockHttpServletResponse = mock(MockHttpServletResponse.class);
+    when(mockMockHttpServletResponse.getHeader(TEST_HEADER_KEY_1)).thenReturn(TEST_HEADER_VALUE_1);
+    when(mockMvcResult.getResponse()).thenReturn(mockMockHttpServletResponse);
+    when(mockResultActions.andReturn()).thenReturn(mockMvcResult);
   }
 
-  @Test
-  void WHEN_answerAsString_THEN_return_expected_value() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString();
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
+  @Nested
+  class answerAsResultActions {
 
-    // Act
-    var result = this.testAnswer.answerAsString();
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsResultActions_THEN_return_resultActions() {
+      // Act
+      var resultActions = testAnswer.answerAsResultActions();
 
-    // Assert
-    assertThat(result, is(TEST_BODY_JSON));
+      // Assert
+      assertThat(resultActions, is(mockResultActions));
+    }
   }
 
-  @Test
-  void GIVEN_exception_WHEN_answerAsString_THEN_throw_exception() throws Exception {
-    // Arrange
-    this.setup.mockThrowGetContentAsString();
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
+  @Nested
+  class answerAsString {
 
-    // Act
-    var exception = assertThrows(Exception.class, () -> this.testAnswer.answerAsString());
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsString_THEN_return_expected_string() {
+      // Arrange
+      useContentAsString(TEST_A1_JSON);
 
-    // Assert
-    assertThat(exception.getClass(), is(TestAnswerException.class));
+      // Act
+      var result = testAnswer.answerAsString();
+
+      // Assert
+      assertThat(result, is(TEST_A1_JSON));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_null_WHEN_answerAsString_THEN_return_nullValue() {
+      // Arrange
+      useContentAsString(null);
+
+      // Act
+      var result = testAnswer.answerAsString();
+
+      // Assert
+      assertThat(result, is(nullValue()));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_exception_WHEN_answerAsString_THEN_throw_TestAnswerException() {
+      // Arrange
+      useContentAsStringException();
+
+      // Act & Assert
+      assertThrows(TestAnswerException.class, () -> testAnswer.answerAsString());
+    }
   }
 
-  @Test
-  void WHEN_answerAsByte_THEN_mvc_getContentAsByteArray_is_called() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsByteArray();
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
+  @Nested
+  class answerAsByte {
 
-    // Act
-    this.testAnswer.answerAsByte();
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsByte_THEN_return_expected_byte() {
+      // Arrange
+      useContentAsByte();
 
-    // Assert
-    this.setup.verifyGetContentAsByteArray();
+      // Act
+      var resultByte = testAnswer.answerAsByte();
+
+      // Assert
+      assertThat(resultByte, is(TEST_BYTE));
+    }
   }
 
-  @Test
-  void WHEN_answerAsByte_THEN_return_expected_value() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsByteArray();
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
+  @Nested
+  class answerAsObject {
 
-    // Act
-    var result = this.testAnswer.answerAsByte();
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsObject_THEN_return_expected_object() {
+      // Arrange
+      useContentAsString(TEST_A1_JSON);
 
-    // Assert
-    assertThat(result, is(TEST_BYTE));
+      // Act
+      var result = testAnswer.answerAsObject(TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(A1));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_null_WHEN_answerAsObject_THEN_return_nullValue() {
+      // Arrange
+      useContentAsString(null);
+
+      // Act
+      var result = testAnswer.answerAsObject(TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(nullValue()));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_exception_WHEN_answerAsObject_THEN_throw_TestAnswerException() {
+      // Arrange
+      useContentAsStringException();
+
+      // Act & Assert
+      assertThrows(
+          TestAnswerException.class, () -> testAnswer.answerAsObject(TestObjectSimple.class));
+    }
+
+    @Test
+    @SneakyThrows
+    @SuppressWarnings("all")
+    void GIVEN_null_as_resultType_WHEN_answerAsObject_THEN_throw_TestAnswerException() {
+      // Act & Assert
+      assertThrows(NullPointerException.class, () -> testAnswer.answerAsObject(null));
+    }
   }
 
-  @Test
-  @SuppressWarnings("ConstantConditions")
-  void GIVEN_null_WHEN_answerAsObject_THEN_throw_NullPointerException() {
-    assertThrows(NullPointerException.class, () -> this.testAnswer.answerAsObject(null));
+  @Nested
+  class answerAsList {
+
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsList_THEN_return_expected_object() {
+      // Arrange
+      useContentAsString(TEST_LIST_A1_A2_JSON);
+
+      // Act
+      var result = testAnswer.answerAsList(TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(TEST_LIST_A1_A2));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_null_WHEN_answerAsList_THEN_return_nullValue() {
+      // Arrange
+      useContentAsString(null);
+
+      // Act
+      var result = testAnswer.answerAsList(TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(nullValue()));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_exception_WHEN_answerAsList_THEN_throw_TestAnswerException() {
+      // Arrange
+      useContentAsStringException();
+
+      // Act & Assert
+      assertThrows(
+          TestAnswerException.class, () -> testAnswer.answerAsList(TestObjectSimple.class));
+    }
+
+    @Test
+    @SneakyThrows
+    @SuppressWarnings("all")
+    void GIVEN_null_as_resultType_WHEN_answerAsList_THEN_throw_TestAnswerException() {
+      // Act & Assert
+      assertThrows(NullPointerException.class, () -> testAnswer.answerAsList(null));
+    }
   }
 
-  @Test
-  void GIVEN_exception_WHEN_answerAsObject_THEN_throw_TestAnswerException() throws Exception {
-    // Arrange
-    when(this.mvc.perform(any())).thenThrow(Exception.class);
+  @Nested
+  class answerAsSet {
 
-    // Act
-    var exception =
-        assertThrows(Exception.class, () -> this.testAnswer.answerAsObject(TestObjectSimple.class));
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsSet_THEN_return_expected_object() {
+      // Arrange
+      useContentAsString(TEST_SET_A1_A2_JSON);
 
-    // Assert
-    assertThat(exception.getClass(), is(TestAnswerException.class));
+      // Act
+      var result = testAnswer.answerAsSet(TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(TEST_SET_A1_A2));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_null_WHEN_answerAsSet_THEN_return_nullValue() {
+      // Arrange
+      useContentAsString(null);
+
+      // Act
+      var result = testAnswer.answerAsSet(TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(nullValue()));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_exception_WHEN_answerAsSet_THEN_throw_TestAnswerException() {
+      // Arrange
+      useContentAsStringException();
+
+      // Act & Assert
+      assertThrows(TestAnswerException.class, () -> testAnswer.answerAsSet(TestObjectSimple.class));
+    }
+
+    @Test
+    @SneakyThrows
+    @SuppressWarnings("all")
+    void GIVEN_null_as_resultType_WHEN_answerAsSet_THEN_throw_TestAnswerException() {
+      // Act & Assert
+      assertThrows(NullPointerException.class, () -> testAnswer.answerAsSet(null));
+    }
   }
 
-  @Test
-  void GIVEN_wrong_class_WHEN_answerAsObject_THEN_throw_Exception() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_A1_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
+  @Nested
+  class answerAsMap {
 
-    // Assert
-    assertThrows(
-        TestAnswerException.class,
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsMap_THEN_return_expected_object() {
+      // Arrange
+      useContentAsString(TEST_MAP_A1_A2_JSON);
 
-        // Act
-        () -> this.testAnswer.answerAsObject(String.class));
+      // Act
+      var result = testAnswer.answerAsMap(Integer.class, TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(TEST_MAP_A1_A2));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_null_WHEN_answerAsMap_THEN_return_nullValue() {
+      // Arrange
+      useContentAsString(null);
+
+      // Act
+      var result = testAnswer.answerAsMap(Integer.class, TestObjectSimple.class);
+
+      // Assert
+      assertThat(result, is(nullValue()));
+    }
+
+    @Test
+    @SneakyThrows
+    void GIVEN_exception_WHEN_answerAsMap_THEN_throw_TestAnswerException() {
+      // Arrange
+      useContentAsStringException();
+
+      // Act & Assert
+      assertThrows(
+          TestAnswerException.class,
+          () -> testAnswer.answerAsMap(Integer.class, TestObjectSimple.class));
+    }
+
+    @Test
+    @SneakyThrows
+    @SuppressWarnings("all")
+    void GIVEN_null_as_keyType_WHEN_answerAsMap_THEN_throw_TestAnswerException() {
+      // Act & Assert
+      assertThrows(
+          NullPointerException.class, () -> testAnswer.answerAsMap(null, TestObjectSimple.class));
+    }
+
+    @Test
+    @SneakyThrows
+    @SuppressWarnings("all")
+    void GIVEN_null_as_valueType_WHEN_answerAsMap_THEN_throw_TestAnswerException() {
+      // Act & Assert
+      assertThrows(NullPointerException.class, () -> testAnswer.answerAsMap(Integer.class, null));
+    }
   }
 
-  @Test
-  void WHEN_answerAsObject_THEN_return_expected_object() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_A1_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
+  @Nested
+  class answerHeader {
 
-    // Act
-    var result = this.testAnswer.answerAsObject(TestObjectSimple.class);
+    @Test
+    @SneakyThrows
+    void WHEN_answerHeader_THEN_return_expected_header_value() {
+      // Arrange
+      useHeader();
 
-    // Assert
-    assertThat(result, is(A1));
+      // Act
+      var resultHeader = testAnswer.answerHeader(TEST_HEADER_KEY_1);
+
+      // Assert
+      assertThat(resultHeader, is(TEST_HEADER_VALUE_1));
+    }
   }
 
-  @Test
-  @SuppressWarnings("ConstantConditions")
-  void GIVEN_null_WHEN_answerAsList_THEN_throw_NullPointerException() {
-    assertThrows(NullPointerException.class, () -> this.testAnswer.answerAsList(null));
-  }
+  @Nested
+  @SuppressWarnings("all")
+  class answerAsVoid {
 
-  @Test
-  void GIVEN_exception_WHEN_answerAsList_THEN_throw_TestAnswerException() throws Exception {
-    // Arrange
-    when(this.mvc.perform(any())).thenThrow(Exception.class);
+    @Test
+    @SneakyThrows
+    void WHEN_answerAsVoid_THEN_verifyNoInteractions() {
+      // Arrange
+      var mockTestRequestStrategy = Mockito.mock(TestRequestBaseStrategy.class);
+      var mockTestRequestStrategyFactory =
+          MockTestRequestStrategyFactory.mockTestRequestStrategyFactory(mockTestRequestStrategy);
 
-    // Act
-    var exception =
-        assertThrows(Exception.class, () -> this.testAnswer.answerAsList(TestObjectSimple.class));
+      when(mockTestRequestStrategy.apply(TEST_REQUEST_DTO)).thenReturn(mockRequestBuilder);
 
-    // Assert
-    assertThat(exception.getClass(), is(TestAnswerException.class));
-  }
+      // Act
+      testAnswer.answerVoid();
 
-  @Test
-  void GIVEN_wrong_class_WHEN_answerAsList_THEN_throw_Exception() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_LIST_A1_A2_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Assert
-    assertThrows(
-        TestAnswerException.class,
-
-        // Act
-        () -> this.testAnswer.answerAsList(String.class));
-  }
-
-  @Test
-  void WHEN_answerAsList_THEN_return_expected_list() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_LIST_A1_A2_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Act
-    var result = this.testAnswer.answerAsList(TestObjectSimple.class);
-
-    // Assert
-    assertThat(result, is(TEST_LIST_A1_A2));
-  }
-
-  @Test
-  @SuppressWarnings("ConstantConditions")
-  void GIVEN_null_WHEN_answerAsSet_THEN_throw_NullPointerException() {
-    assertThrows(NullPointerException.class, () -> this.testAnswer.answerAsSet(null));
-  }
-
-  @Test
-  void GIVEN_exception_WHEN_answerAsSet_THEN_throw_TestAnswerException() throws Exception {
-    // Arrange
-    when(this.mvc.perform(any())).thenThrow(Exception.class);
-
-    // Act
-    var exception =
-        assertThrows(Exception.class, () -> this.testAnswer.answerAsSet(TestObjectSimple.class));
-
-    // Assert
-    assertThat(exception.getClass(), is(TestAnswerException.class));
-  }
-
-  @Test
-  void GIVEN_wrong_class_WHEN_answerAsSet_THEN_throw_Exception() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_LIST_A1_A2_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Assert
-    assertThrows(
-        TestAnswerException.class,
-
-        // Act
-        () -> this.testAnswer.answerAsSet(String.class));
-  }
-
-  @Test
-  void WHEN_answerAsSet_THEN_return_expected_set() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_SET_A1_A2_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Act
-    var result = this.testAnswer.answerAsSet(TestObjectSimple.class);
-
-    // Assert
-    assertThat(result, is(TEST_SET_A1_A2));
-  }
-
-  @ParameterizedTest
-  @MethodSource("provideNullParameters")
-  @SuppressWarnings("ConstantConditions")
-  void GIVEN_provideNullParameters_WHEN_answerAsMap_THEN_throw_NullPointerException(
-      Class<String> targetKeyClass, Class<String> targetValueClass) {
-
-    assertThrows(
-        NullPointerException.class,
-        () -> this.testAnswer.answerAsMap(targetKeyClass, targetValueClass));
-  }
-
-  private static Stream<Arguments> provideNullParameters() {
-    return Stream.of(
-        Arguments.of(null, String.class),
-        Arguments.of(String.class, null),
-        Arguments.of(null, null));
-  }
-
-  @Test
-  void GIVEN_exception_WHEN_answerAsMap_THEN_throw_TestAnswerException() throws Exception {
-    // Arrange
-    when(this.mvc.perform(any())).thenThrow(Exception.class);
-
-    // Act
-    var exception =
-        assertThrows(
-            Exception.class,
-            () -> this.testAnswer.answerAsMap(Boolean.class, TestObjectSimple.class));
-
-    // Assert
-    assertThat(exception.getClass(), is(TestAnswerException.class));
-  }
-
-  @ParameterizedTest
-  @MethodSource("wrongClasses")
-  void GIVEN_wrongClasses_WHEN_answerAsMap_THEN_throw_Exception(
-      Class<?> targetKeyClass, Class<?> targetValueClass) throws Exception {
-
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_LIST_A1_A2_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Assert
-    assertThrows(
-        TestAnswerException.class,
-
-        // Act
-        () -> this.testAnswer.answerAsMap(targetKeyClass, targetValueClass));
-  }
-
-  private static Stream<Arguments> wrongClasses() {
-    return Stream.of(
-        Arguments.of(Boolean.class, String.class),
-        Arguments.of(String.class, TestObjectSimple.class),
-        Arguments.of(String.class, String.class));
-  }
-
-  @Test
-  void WHEN_answerAsMap_THEN_return_expected_map() throws Exception {
-    // Arrange
-    this.setup.mockGetContentAsString(TEST_MAP_A1_A2_JSON);
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Act
-    var result = this.testAnswer.answerAsMap(Integer.class, TestObjectSimple.class);
-
-    // Assert
-    assertThat(result, is(TEST_MAP_A1_A2));
-  }
-
-  @Test
-  void WHEN_answerHeader_THEN_mvc_getContentAsByteArray_is_called() throws Exception {
-    // Arrange
-    this.setup.mockGetHeader();
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Act
-    this.testAnswer.answerHeader(TEST_HEADER_KEY_1);
-
-    // Assert
-    this.setup.verifyGetHeader();
-  }
-
-  @Test
-  void WHEN_answerHeader_THEN_return_expected_value() throws Exception {
-    // Arrange
-    this.setup.mockGetHeader();
-    this.testAnswer = new TestAnswerImpl(this.testRequestContext, this.mockRequestBuilder);
-
-    // Act
-    var result = this.testAnswer.answerHeader(TEST_HEADER_KEY_1);
-
-    // Assert
-    assertThat(result, is(TEST_HEADER_VALUE_1));
+      // Assert
+      verify(testRequestBean.mvc()).perform(mockRequestBuilder);
+      mockTestRequestStrategyFactory.close();
+    }
   }
 }